Split models.py into models folder
