+++ publish/models/__init__.py
@@ -1,0 +1,3 @@
+from .asset import Asset
+from .dandiset import Dandiset
+from .version import Version
